+++ tests/Unit/Services/Contact/Conversation/CreateConversationTest.php
@@ -16,7 +16,8 @@ class CreateConversationTest extends TestCase
 {
     use DatabaseTransactions;
 
-    public function test_it_stores_a_conversation()
+    /** @test */
+    public function it_stores_a_conversation()
     {
         $contact = factory(Contact::class)->create([]);
         $contactFieldType = factory(ContactFieldType::class)->create([
@@ -45,7 +46,8 @@ class CreateConversationTest extends TestCase
         );
     }
 
-    public function test_it_fails_if_wrong_parameters_are_given()
+    /** @test */
+    public function it_fails_if_wrong_parameters_are_given()
     {
         $contact = factory(Contact::class)->create([]);
 
@@ -59,7 +61,8 @@ class CreateConversationTest extends TestCase
         app(CreateConversation::class)->execute($request);
     }
 
-    public function test_it_throws_an_exception_if_contact_is_not_linked_to_account()
+    /** @test */
+    public function it_throws_an_exception_if_contact_is_not_linked_to_account()
     {
         $account = factory(Account::class)->create();
         $contact = factory(Contact::class)->create();
@@ -79,7 +82,8 @@ class CreateConversationTest extends TestCase
         app(CreateConversation::class)->execute($request);
     }
 
-    public function test_it_throws_an_exception_if_contactfieldtype_is_not_linked_to_account()
+    /** @test */
+    public function it_throws_an_exception_if_contactfieldtype_is_not_linked_to_account()
     {
         $contact = factory(Contact::class)->create([]);
         $contactFieldType = factory(ContactFieldType::class)->create([]);